Rename `pv service` to `pv daemon` to free up `service` namespace
The `service` subcommand namespace will be used for managing external
services like MySQL, PostgreSQL, Docker containers, etc. The daemon
management (launchd auto-start) moves to `pv daemon install/uninstall`.
